// src/types/common/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  applicants: {
    Tables: {
// ...
      applicant_requirement_media: {
        Row: {
          alt_text: string | null
          applicant_requirement_id: number | null
          created_at: string
          description: string | null
          filename: string | null
          id: number
          mime_type: string | null
          path: string | null
          profiles_id: string | null
          size: number | null
          updated_at: string | null
        }
        Insert: {
          alt_text?: string | null
          applicant_requirement_id?: number | null
          created_at?: string
          description?: string | null
          filename?: string | null
          id?: never
          mime_type?: string | null
          path?: string | null
          profiles_id?: string | null
          size?: number | null
          updated_at?: string | null
        }
        Update: {
          alt_text?: string | null
          applicant_requirement_id?: number | null
          created_at?: string
          description?: string | null
          filename?: string | null
          id?: never
          mime_type?: string | null
          path?: string | null
          profiles_id?: string | null
          size?: number | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "applicant_requirement_media_applicant_requirement_id_fkey"
            columns: ["applicant_requirement_id"]
            isOneToOne: false
            referencedRelation: "applicant_requirements"
            referencedColumns: ["id"]
          },
        ]
      }
      applicant_requirements: {
        Row: {
          created_at: string
          id: number
          profile_id: string | null
          remarks: string | null
          requirement_id: number | null
          status: Database["public"]["Enums"]["status_type"] | null
        }
        Insert: {
          created_at?: string
          id?: never
          profile_id?: string | null
          remarks?: string | null
          requirement_id?: number | null
          status?: Database["public"]["Enums"]["status_type"] | null
        }
        Update: {
          created_at?: string
          id?: never
          profile_id?: string | null
          remarks?: string | null
          requirement_id?: number | null
          status?: Database["public"]["Enums"]["status_type"] | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  core: {
    Tables: {
// ...
      profiles: {
        Row: {
          barangay: string | null
          birthdate: string | null
          contact_number: string | null
          created_at: string
          firstname: string | null
          gender: Database["core"]["Enums"]["gender_type"] | null
          geographic: string | null
          id: string
          is_4ps: boolean | null
          is_pwd: boolean | null
          last_login: string | null
          lastname: string | null
          middlename: string | null
          province: string | null
          region: string | null
          role: Database["core"]["Enums"]["user_role"] | null
          status: Database["core"]["Enums"]["status_type"] | null
          updated_at: string | null
          username: string | null
        }
        Insert: {
          barangay?: string | null
          birthdate?: string | null
          contact_number?: string | null
          created_at?: string
          firstname?: string | null
          gender?: Database["core"]["Enums"]["gender_type"] | null
          geographic?: string | null
          id?: string
          is_4ps?: boolean | null
          is_pwd?: boolean | null
          last_login?: string | null
          lastname?: string | null
          middlename?: string | null
          province?: string | null
          region?: string | null
          role?: Database["core"]["Enums"]["user_role"] | null
          status?: Database["core"]["Enums"]["status_type"] | null
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          barangay?: string | null
          birthdate?: string | null
          contact_number?: string | null
          created_at?: string
          firstname?: string | null
          gender?: Database["core"]["Enums"]["gender_type"] | null
          geographic?: string | null
          id?: string
          is_4ps?: boolean | null
          is_pwd?: boolean | null
          last_login?: string | null
          lastname?: string | null
          middlename?: string | null
          province?: string | null
          region?: string | null
          role?: Database["core"]["Enums"]["user_role"] | null
          status?: Database["core"]["Enums"]["status_type"] | null
          updated_at?: string | null
          username?: string | null
        }
        Relationships: []
      }
    }
// ...
  }
  employers: {
    Tables: {
      companies: {
        Row: {
          business_type: string | null
          company_address: string | null
          company_contact: string | null
          company_description: string | null
          company_email: string | null
          company_name: string | null
          created_at: string
          employee_count: number | null
          id: number
          industry: string | null
          latitude: number | null
          longitude: number | null
          profile_id: string | null
          registration_number: string | null
          updated_at: string | null
          verification_status: Database["public"]["Enums"]["status_type"] | null
          website: string | null
        }
        Insert: {
          business_type?: string | null
          company_address?: string | null
          company_contact?: string | null
          company_description?: string | null
          company_email?: string | null
          company_name?: string | null
          created_at?: string
          employee_count?: number | null
          id?: never
          industry?: string | null
          latitude?: number | null
          longitude?: number | null
          profile_id?: string | null
          registration_number?: string | null
          updated_at?: string | null
          verification_status?:
            | Database["public"]["Enums"]["status_type"]
            | null
          website?: string | null
        }
        Update: {
          business_type?: string | null
          company_address?: string | null
          company_contact?: string | null
          company_description?: string | null
          company_email?: string | null
          company_name?: string | null
          created_at?: string
          employee_count?: number | null
          id?: never
          industry?: string | null
          latitude?: number | null
          longitude?: number | null
          profile_id?: string | null
          registration_number?: string | null
          updated_at?: string | null
          verification_status?:
            | Database["public"]["Enums"]["status_type"]
            | null
          website?: string | null
        }
        Relationships: []
      }
      company_members: {
        Row: {
          company_id: number | null
          created_at: string | null
          id: number
          profile_id: string | null
          role: Database["public"]["Enums"]["user_role"]
          status: Database["public"]["Enums"]["status_type"] | null
          updated_at: string | null
        }
        Insert: {
          company_id?: number | null
          created_at?: string | null
          id?: never
          profile_id?: string | null
          role?: Database["public"]["Enums"]["user_role"]
          status?: Database["public"]["Enums"]["status_type"] | null
          updated_at?: string | null
        }
        Update: {
          company_id?: number | null
          created_at?: string | null
          id?: never
          profile_id?: string | null
          role?: Database["public"]["Enums"]["user_role"]
          status?: Database["public"]["Enums"]["status_type"] | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "company_members_company_id_fkey"
            columns: ["company_id"]
            isOneToOne: false
            referencedRelation: "companies"
            referencedColumns: ["id"]
          },
        ]
      }
      employer_requirement_media: {
        Row: {
          alt_text: string | null
          created_at: string
          description: string | null
          employer_requirement_id: number | null
          filename: string | null
          id: number
          mime_type: string | null
          path: string | null
          profile_id: string | null
          size: number | null
          updated_at: string | null
        }
        Insert: {
          alt_text?: string | null
          created_at?: string
          description?: string | null
          employer_requirement_id?: number | null
          filename?: string | null
          id?: never
          mime_type?: string | null
          path?: string | null
          profile_id?: string | null
          size?: number | null
          updated_at?: string | null
        }
        Update: {
          alt_text?: string | null
          created_at?: string
          description?: string | null
          employer_requirement_id?: number | null
          filename?: string | null
          id?: never
          mime_type?: string | null
          path?: string | null
          profile_id?: string | null
          size?: number | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "employer_requirement_media_employer_requirement_id_fkey"
            columns: ["employer_requirement_id"]
            isOneToOne: false
            referencedRelation: "employer_requirements"
            referencedColumns: ["id"]
          },
        ]
      }
      employer_requirements: {
        Row: {
          employer_id: number | null
          id: number
          remarks: string | null
          requirement_id: number | null
          status: Database["public"]["Enums"]["status_type"] | null
          submitted_at: string
        }
        Insert: {
          employer_id?: number | null
          id?: never
          remarks?: string | null
          requirement_id?: number | null
          status?: Database["public"]["Enums"]["status_type"] | null
          submitted_at?: string
        }
        Update: {
          employer_id?: number | null
          id?: never
          remarks?: string | null
          requirement_id?: number | null
          status?: Database["public"]["Enums"]["status_type"] | null
          submitted_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "employer_requirements_employer_id_fkey"
            columns: ["employer_id"]
            isOneToOne: false
            referencedRelation: "companies"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
  public: {
    Tables: {
// ...
      requirement_templates: {
        Row: {
          created_at: string
          description: string | null
          id: number
          name: string | null
          requirement_type: string | null
          updated_at: string | null
        }
        Insert: {
          created_at?: string
          description?: string | null
          id?: number
          name?: string | null
          requirement_type?: string | null
          updated_at?: string | null
        }
        Update: {
          created_at?: string
          description?: string | null
          id?: number
          name?: string | null
          requirement_type?: string | null
          updated_at?: string | null
        }
        Relationships: []
      }
    }
// ...
  }
// ...
}